[TensorExpr] Add some graph-rewrite passes to prepare models for AOT compilation.
These passes should not be used generally as they change API of the
model's forward method, but they help experimenting with the model and
ironing out all the kinks before it can be compiled properly. In the
long run ideally we should provide a better way to enable such
experiments.

ghstack-source-id: a4d76aaca6ed85a8e1dd29aea445bdafe7f9e338
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66515